refactor: 예외를 별도의 패키지로 이동
